backend/requirements.txt
fastapi
uvicorn
sqlalchemy
python-dotenv
snowflake-sqlalchemy
boto3
python-multipart